Backend/requirements.txt
advanced_alchemy==0.3.4
alembic==1.12.0
annotated-types==0.6.0
anyio==4.0.0
certifi==2023.7.22
click==8.1.7
EditorConfig==0.12.3
Faker==19.11.0
fast-query-parsers==1.0.3
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
jsbeautifier==1.14.9
litestar==2.2.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgspec==0.18.4
multidict==6.0.4
polyfactory==2.10.0
psycopg==3.1.12
psycopg-binary==3.1.12
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
rich==13.6.0
rich-click==1.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
uuid7==0.1.0
uvicorn==0.23.2
# uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
